Clean up the error module
Formerly errors. Can now also convert from Record.
