+++ Cargo.toml
@@ -22,8 +22,10 @@ default-target = "x86_64-pc-windows-msvc"
 name = "deferred"
 crate-type = ["cdylib"]
 path = "examples/deferred/lib.rs"
+required-features = ["nightly"]
 
 [[example]]
 name = "skip"
 crate-type = ["cdylib"]
 path = "examples/skip/lib.rs"
+required-features = ["nightly"]
