+++ src/error/mod.rs
@@ -5,6 +5,8 @@ use crate::Record;
 use std::fmt::Display;
 use std::num::{NonZeroU32, TryFromIntError};
 
+pub mod experimental;
+
 /// Results returned by this crate.
 pub type Result<T> = std::result::Result<T, Error>;
 
@@ -118,6 +120,12 @@ impl From<std::string::FromUtf8Error> for Error {
     }
 }
 
+impl From<Record> for Error {
+    fn from(record: Record) -> Self {
+        Error::from_error_record(record)
+    }
+}
+
 #[derive(Debug)]
 enum Context {
     Simple(ErrorKind),
@@ -129,141 +137,6 @@ enum Context {
 struct Custom {
     kind: ErrorKind,
     error: Box<dyn std::error::Error + Send + Sync>,
-}
-
-#[cfg(feature = "nightly")]
-pub mod experimental {
-    use super::{Error, ErrorKind};
-    use crate::ffi;
-    use std::convert::Infallible;
-    use std::fmt::Display;
-    use std::num::NonZeroU32;
-    use std::ops::{ControlFlow, FromResidual, Try};
-
-    /// A result to return from a custom action.
-    ///
-    /// This allows you to use the `?` operator to map any `Result<T, E>` to [`CustomActionResult::Fail`].
-    ///
-    /// # Example
-    ///
-    /// ```no_run
-    /// use std::ffi::OsString;
-    /// use msica::{Session, CustomActionResult};
-    ///
-    /// #[no_mangle]
-    /// pub extern "C" fn MyCustomAction(session: Session) -> CustomActionResult {
-    ///     let productName = session.property("ProductName")?;
-    ///
-    ///     // Do something with `productName`.
-    ///
-    ///     CustomActionResult::Succeed
-    /// }
-    /// ```
-    #[derive(Debug, Copy, Clone, Eq, PartialEq)]
-    #[repr(u32)]
-    pub enum CustomActionResult {
-        /// Completed actions successfully.
-        Succeed = ffi::ERROR_SUCCESS,
-
-        /// Skip remaining actions. Not an error.
-        Skip = ffi::ERROR_NO_MORE_ITEMS,
-
-        /// User terminated prematurely.
-        Cancel = ffi::ERROR_INSTALL_USEREXIT,
-
-        /// Unrecoverable error occurred.
-        Fail = ffi::ERROR_INSTALL_FAILURE,
-
-        /// Action not executed.
-        NotExecuted = ffi::ERROR_FUNCTION_NOT_CALLED,
-    }
-
-    impl Display for CustomActionResult {
-        fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-            let error = match &self {
-                Self::Succeed => "succeeded",
-                Self::Skip => "skip remaining actions",
-                Self::Cancel => "user canceled",
-                Self::Fail => "failed",
-                Self::NotExecuted => "not executed",
-            };
-
-            write!(f, "{}", error)
-        }
-    }
-
-    impl From<u32> for CustomActionResult {
-        fn from(code: u32) -> Self {
-            match code {
-                ffi::ERROR_SUCCESS => CustomActionResult::Succeed,
-                ffi::ERROR_NO_MORE_ITEMS => CustomActionResult::Skip,
-                ffi::ERROR_INSTALL_USEREXIT => CustomActionResult::Cancel,
-                ffi::ERROR_FUNCTION_NOT_CALLED => CustomActionResult::NotExecuted,
-                _ => CustomActionResult::Fail,
-            }
-        }
-    }
-
-    #[allow(clippy::from_over_into)]
-    impl Into<u32> for CustomActionResult {
-        fn into(self) -> u32 {
-            self as u32
-        }
-    }
-
-    /// This type is an implementation detail and not intended for direct use.
-    #[doc(hidden)]
-    pub struct CustomActionResultCode(NonZeroU32);
-
-    impl From<CustomActionResult> for CustomActionResultCode {
-        fn from(value: CustomActionResult) -> Self {
-            CustomActionResultCode(NonZeroU32::new(value.into()).unwrap())
-        }
-    }
-
-    impl Try for CustomActionResult {
-        type Output = u32;
-        type Residual = CustomActionResultCode;
-
-        fn branch(self) -> ControlFlow<Self::Residual, Self::Output> {
-            match self {
-                Self::Succeed => ControlFlow::Continue(ffi::ERROR_SUCCESS),
-                _ => ControlFlow::Break(self.into()),
-            }
-        }
-
-        fn from_output(_: Self::Output) -> Self {
-            CustomActionResult::Succeed
-        }
-    }
-
-    impl FromResidual for CustomActionResult {
-        fn from_residual(residual: CustomActionResultCode) -> Self {
-            match residual.0.into() {
-                ffi::ERROR_NO_MORE_ITEMS => CustomActionResult::Skip,
-                ffi::ERROR_INSTALL_USEREXIT => CustomActionResult::Cancel,
-                ffi::ERROR_INSTALL_FAILURE => CustomActionResult::Fail,
-                ffi::ERROR_FUNCTION_NOT_CALLED => CustomActionResult::NotExecuted,
-                code => panic!("unexpected error code {}", code),
-            }
-        }
-    }
-
-    impl FromResidual<Result<Infallible, Error>> for CustomActionResult {
-        fn from_residual(residual: Result<Infallible, Error>) -> Self {
-            let error = residual.unwrap_err();
-            match error.kind() {
-                ErrorKind::ErrorCode(code) => CustomActionResult::from(code.get()),
-                _ => CustomActionResult::Fail,
-            }
-        }
-    }
-
-    impl<E: std::error::Error> FromResidual<std::result::Result<Infallible, E>> for CustomActionResult {
-        default fn from_residual(_: std::result::Result<Infallible, E>) -> Self {
-            CustomActionResult::Fail
-        }
-    }
 }
 
 #[cfg(test)]
@@ -289,7 +162,7 @@ mod tests {
             vec![Field::StringData("text".to_owned())],
         )
         .expect("failed to create record");
-        let error = Error::from_error_record(record);
+        let error: Error = record.into();
         assert_eq!(&ErrorKind::ErrorRecord, error.kind());
         assert_eq!("error text", error.to_string());
     }
